Move `<CardFooter />` to `card-footer/` subfolder
