+++ packages/components/src/card/index.js
@@ -1,3 +1,4 @@
 export { default as Card } from './card';
 export { default as CardBody } from './card-body';
 export { default as CardDivider } from './card-divider';
+export { default as CardFooter } from './card-footer';
